python/co2-emission/requirements.txt
scramjet-framework-py
aiohttp